keyserver: reduce config post-processing
When we were using configuration files, we needed multiple stages to
ensure that valid data was provided, and to cross-reference different
parts of the configuration file. This is no longer necessary, now that
we generate the configuration directly, and can produce it already
validated and cross-referenced.

Eliminate some of the processing stages.