auth-monitor: eliminate machine.list
There's no reason to pull down machine.list from the keyserver when the
setup.yaml file already present on the supervisor contains the same
info. This means we can eliminate the entire machine.list file from the
keysystem. Pull down a different file instead in the auth-monitor so
that we still accomplish the secondary goal of confirming that the
keyclient is serving static files correctly.