checklist: remove duplicate effort requirement
At this point, since we have CI, it is no longer necessary to manually
test cluster deployment for each change -- or even build it locally.
The only manual testing that's actually *needed* is to confirm that the
newly added functionality works correctly.

Also rearrange entries such that CI and code review are explicitly
noted on the checklist.