keyinitadmit: load local hostname from setup.yaml
We want to move away from using a configuration uploaded by spire,
which means parsing the cluster's setup.yaml directly. Add parsing for
the setup.yaml format. Use this parsing to remove the hostname
parameter from keyinitadmit that can be computed from setup.yaml.

The reason for this commit is the inclusion of the setup.yaml format
parser, but the actual primary change is the change to keyinitadmit.