keyserver: generate https certificate in place
Currently, we're hard-coding the name of the supervisor node into the
keyserver certificate that's distributed to every node in the cluster.
This is problematic, because it would make it hard to change the name
of the supervisor without a very careful replacement process.

Generate a HTTPS certificate on the fly for the current supervisor node
each time the keyserver starts up, so the supervisor name doesn't need
to be in any long-lived certificate, and renew it whenever it expires.
Because we can reuse the existing cluster HTTPS CA to sign this
certificate, this also means that we decrease the number of long-lived
keys we have to manage, which is good for simplicity.

We can trust the cluster CA about as much as we trust the keyserver
certificate, so replacing our verifications against the keyserver
certificate with verifications against the CA doesn't decrease the
security of the cluster by any significant amount.