keyinitadmit: remove duplicate config parse
We're already loading data about all the supervisors into memory; we
shouldn't need to grab the name of the supervisor with a separate
sequence. Add an accessor to get the supervisor node metadata, and
store the supervisor node's DNS address so that we can use it in
keyinitadmit.

This also prepares us for changing how the keyserver certificate is
generated in the next commit.