keysystem: fix certificate authority naming
First, calling the cluster HTTPS CA "clustertls" is inspecific; clarify
that it is a certificate authority in the usual sense (i.e. used to
sign HTTPS webpage certificates.)

Second, adjust all of the authority key and certificate filenames to
"NAME.key" and "NAME.pem" for TLS authorities and "NAME" and "NAME.pub"
for SSH authorities.