keysystem: simplify authority description model
Previously, we described authorities in the configuration with four
settings: type (TLS or SSH or static), keyfile name, certfile name, and
DNS name list. We do not need all of this configurability, and it makes
authority handling more complex than it needs to be.

Replace this with two choices: SSH or TLS, and name. Derive everything
else from these. Note that the removal of the static authority option
occurred in the last commit, but this commit takes advantage of it to
turn the option into a boolean. Note also that nothing has used the DNS
name list since the elimination of the long-lived keyserver HTTPS
certificate.