keyserver: eliminate static authorities
Static authorities were only used in cases where private keys had to be
directly downloaded, and were otherwise generated and handled the same
as TLS authorities. Merge these two concepts by adding the "can have
private keys downloaded with an appropriate privilege" capability to
the TLS authority structure and eliminating the static authority
structure.

This is not a security concern, because the same special privilege
scoped to a particular authority is still needed to download private
keys.