spire: add short hostnames to dns bootstrap
Kubelet uses os.Hostname() by default to determine the hostname it reports
to the API server. This is not the FQDN hostname, so before this commit
Kubernetes was unable to resolve hostnames when performing API calls.
This adds short hostnames to /etc/hosts during DNS bootstrap to allow
Kubernetes to resolve them.

This also auto-generates the other necessary lines for /etc/hosts, since
we currently don't have a reason to have a separate DNS bootstrap step.