crio: assert that crio.conf does not change
When CRI-O changes its configuration file defaults, we need to adjust
the configuration file template we use. As things stand, it's easy to
forget to do this.

To fix this, keep track of the last default config file we've updated
against, and add an assertion to the CRI-O build process so that, when
it changes upstream, the build will fail until we update our copy.

Of course, this requires that we remember to also update crio.conf.in
when we update crio.conf.upstream, but that's easier than remembering
that this step is necessary each time CRI-O gets updated.