kube-state-metrics: eliminate build indirection
Previously, we needed to turn on static linking of kube-state-metrics
in order to use it inside a container. This meant we had to have our
own compilation step to build the binary, and had to patch the bazel
build file to expose the precursors to the binary so that we could do
so.

We no longer turn on static linking, because we no longer use
kube-state-metrics in a container, so we can eliminate this layer of
indirection and just use the binary built by the autogenerated bazel
build scripts for kube-state-metrics directly.