keysystem: populate names and groups for RBAC
The RBAC authorization mode in Kubernetes requires the authentication
system to provide names and groups for everyone who is authenticated;
we previously constructed client certificates without regard to how
they would be processed by Kubernetes, because we weren't doing
anything for which this mattered.

Revise all instances where a kubernetes client is granted to specify
a kubernetes username and possibly a set of groups, as necessary. Note
that groups are provided as "Organizations" in the certificate.

Authentication information provided by this commit:
 - Admins use root:(principal) or root:direct with the builtin cluster
   group system:masters for complete access to everything.
 - API servers use apiserver:(hostname) for their user. This isn't
   necessary, because they never use them to authenticate anywhere, but
   now the naming should be clearer.
 - Kubelets use system:node:(hostname) for their user and system:nodes
   as their only group. This is the standard format that Kubernetes
   expects for the use of the Node authorizer and NodeRestriction
   admission plugin.
 - End users use user:(username) for their user, and have no groups at
   the moment. The username is their kerberos username as extracted
   from their MIT client certificates.