upload: include seqid for key export
The GPG export commands that run under bazel to export public keys for
the developer's signing key will not re-run if the key changes, such as
to extend its expiration date. Deal with this by allowing the developer
to specify a "sequence id". Each time this ID changes, bazel will
re-run the export, so if a developer makes a change like extending the
expiration date of their signing key, they can increment the sequence
ID, and get the new version of their key included.