spire: fix update-registry to pull newer images
Previously, update-registry would upgrade the version of the downloader
package, but wouldn't upgrade the versions of the packages that contain
the actual images. This means that updating a container and trying to
redeploy it with this command wouldn't work. Add a step that upgrades
all other packages, so that new image versions are imported correctly.