kubernetes: port service launcher to go
Rather than using the current mess of shell scripts, use a Go program
to dispatch all of the kubernetes components that need to be launched.

Notably, this allows us to generate a kubeconfig for each component
that needs it, rather than generating it when launching the kubelet and
then assuming that the kubelet was already launched in every other
script.

This also allows us to start specifying some of the key and certificate
file paths as constants, so that they can't get out of sync between the
keysystem and the kubernetes launchers.