kubernetes: support runtime verbosity config
Previously, with the bash scripts used for launching kubernetes
components, it was easy to turn on higher levels of output verbosity
for debugging by editing the scripts to add a --v=N parameter. Now that
these are in Go, it's harder to edit the command lines on the fly.

Allow admins to specify a different verbosity level for the kubernetes
components by populating /etc/homeworld/config/verbosity on the
servers, with a command like:

    echo 3 >/etc/homeworld/config/verbosity