ctrlmgr: configure to use service accounts
The kubernetes controller manager normally uses its access to directly
perform the functions of its different controllers. To be compatible
with the default RBAC configuration, it needs to use service accounts
for each of its different control loops. Enable that option, so that
when we turn on RBAC, the controller manager keeps working.