dns-addon: use standard service account
The default Kubernetes RBAC rules provide the dns-addon in the
kube-system namespace with appropriate permissions to access the
cluster to the extent that it needs to. In order to use these
permissions, it needs to use the correctly named service account.

Configure the dns addon to use the correct service account, so that it
will continue to work when RBAC is enabled.