kubernetes: enable RBAC and Node authorization
In order to protect the cluster from malicious or careless users, we
need to limit the access that everyone has to the system. RBAC is the
most powerful access control tool built into Kubernetes, so we enable
that here.

We also need to provide additional access for the nodes to function
correctly; in previous versions of Kubernetes, broad node access was
included in the default RBAC configuration, but this is no longer the
case; now there is a second authorizer, the Node authorizer, which
grants a node exactly the access it needs to perform work with the pods
that it has. We enable that too.

As well, there is the NodeRestriction admission controller that helps
ensure that nodes authorized with the Node authorizer don't misbehave
in more subtle ways; enable this controller as well.