user-grant: create namespaces and grant access
When a user receives access to the cluster via user-grant, we want to
make sure that they have a namespace where they can deploy their
kubernetes workloads. We can create these automatically in user-grant
before we give them their kubectl key.

As well, now that we use RBAC, users need to be granted the ability to
administrate their own namespaces. We can also set up a default
rolebinding to give them this capability.

In order for user-grant to perform these actions automatically, it
needs special access; grant the access it needs to create new
namespaces and assign users rights to them.

Note that this is broader than we'd like; it would, for example, allow
the user-grant container to grant access to the kube-system namespace,
were it to be compromised. We already have issue 344 filed to deal with
this concern, which already exists due to user-grant's ability to sign
arbitrary kubernetes client certificates.