spire: use yaml-aware templating for kube specs
The existing templating system uses text-based substitution, which
could potentially produce incorrect results unless carefully escaped.

Solve this by performing templating based on parsing the yaml document
and iterating over the branches of the structure. This also changes the
format of the templating syntax from '{{KEY}}' to '{KEY}' to make the
double-quoting introduced by the use of the bazel templating system
less ridiculous.