spire: improve deploy spec export commands
The previous mechanism of exporting kubernetes specs from spire for
inspection had trouble with user-grant, because it required additional
key/value pairs not known in the configuration module.

Change this export process to be a simple option (--export) on the
deploy commands, so that it can be consistent with the normal spec
deployment process.