user-grant: parameterize upstream cert and email
The existing settings for the upstream authentication certificate
authority and for the email domain used by user-grant were hard-coded
into the kubernetes spec, limiting it to use with regular MIT client
certificates.

For the sake of being usable by people besides us, and for the sake of
making this testable in environments where we don't have MIT client
certificates, allow these options to be configured via the setup.yaml
and via storing the upstream certificate in the project directory.