spire: support generating upstream CA for testing
So far, user-grant has been designed for use with MIT's client cert
authentication system, which has limited our ability to automatically
test it, because we couldn't embed a valid MIT cert into the
repository. Make it possible to generate our own CA authority that is
compatible with the MIT CA for our purposes, and issue a client
certificate for a sample user with it.