spire: verify user-grant functionality
We want the ability to confirm automatically that user-grant works
properly, because then we can incorporate it into our standard
integration tests run on Jenkins.

We can't do this if we don't have a client certificate, which we won't
for the MIT CA, but in the case where we generated our own client
certificate authority, and issued such a certificate, we can run a test
just fine.

Add a command to use the locally-issued certificate to authenticate to
the user-grant service, request a kubeconfig file, confirm that it
works, and confirm that user-grant created a rolebinding to grant the
user access. Incorporate this command into the standard testing
sequence when we can.

For the time being, the lack of load balancer support means that we
can't directly access the user-grant service from the machine running
spire, but we can SSH into the master node and perform the request from
there as a stop-gap solution.