jenkins: include user-grant in tests
We want to test user-grant's functionality as part of the standard
integration tests. To do so, we need to generate the appropriate keys,
so that we have a fake HTTPS certificate for the user-grant server and
so that we have a fake certificate authority to issue a client
certificate for testing.