autodeploy: boot ISO as hard drive
On real hardware, the USB drive we insert is interpreted as a hard
drive, which causes the installer to behave differently from how it
behaves in our test environment. Adjust our test environment to more
closely represent the production environment, by inserting the ISO as
a raw hard disk image instead of a CDROM image. Support switching back
to using a CDROM image as an option.