installer: select non-USB drive for grub and root
On real hardware, we install debian with a USB drive, which shows up as
a hard drive to the installer. This messes up both disk selection and
grub installation, because disk selection is only automatically
performed if there's only one disk, and grub installation is always
performed on the first disk. We have two disks, the second of which is
usually the disk that we want.

To solve this, we need to automatically determine which disk we want to
use in a preseed hook, so that we can update the grub install path and
the partitioning target appropriately.

We determine the disk to use by eliminating the disk containing a
partition with the label ISOIMAGE, and hoping that only one disk
remains. This script does not currently support more complex storage
architectures, though we will need to do so to support Ceph.