installer: reconfigure network to use vlans
In order to support deployment environments where network access is
through VLAN trunks, we need to be able to configure the network
appropriately during installation. The debian installer does not
presently support VLAN trunks, so we need to hack in support ourselves.

Replace the netcfg postinstall script so that we can amend the active
and saved network configurations appropriately.