spire: support configuring VLANs for autoinstall
To better simulate the real deployment environment, we need to support
the situation where the VM network is a VLAN trunk. We can simulate
this by attaching the host IP through a vlan interface.

Support configuration of the exact VLAN number to use. If not
specified, use existing behavior of simulating deployment on a
non-trunked interface.