spire: verify that systemd services are running
Most of our verification checks confirm that services are working
indirectly, by their behavior, rather than directly, by checking that
the services are running. The former is a better check that things are
working, but can provide less useful information for debugging when a
service isn't working correctly.

Add a check that confirms that each systemd service is running exactly
where it's supposed to be running, so that, if a service is broken so
much that it can't even start, we can identify that without manually
checking the states of the services on the nodes.

This should also facilitate debugging problems in Jenkins that could be
caused by services not running when they're supposed to be running,
because it's much harder to manually inspect the state of a cluster
deployed as part of Jenkins testing.