spire: use jsonschema for setup.yaml validation
Rather than manually validating the format of setup.yaml while parsing
it, validate the format with jsonschema, so that we can just assume
it's in the correct format.

For the time being, the same validations are used as before, which
means that we aren't, for example, using regexes to validate the format
of each individual configuration entry.