build-chroot: upgrade to buster
 - Remove transitional realpath package (replaced by coreutils)
 - Replace transitional btrfs-tools package with btrfs-progs
 - Replace unsupported openjdk-8-jdk with supported openjdk-11-jdk
 - Add libtinfo5 because it's now needed for container fakeroot setup
 - Add perl-openssl-defaults and dbus-user-session because they seem to
   be required for the chroot to install, though it's not quite clear
   why.
 - Add libbtrfs-dev, because it needs to be pulled in explicitly for
   the CRI-O build now.

Note that this commit will not build correctly at this point, because
using a buster chroot means the fakeroot/fakechroot invocation used to
set up debian containers will not have the correct host libraries. As
well, a buggy version of debootstrap is currently in the debian repo,
which also breaks the debian container scripts. The next two commits
will be required for a successful build.