debian: bump containers to latest buster
Stretch is an entire release behind -- and we're over a year out of
date. Get us to the latest version of the latest release.

This makes it possible to build again, because the build chroot now
matches the container release closely enough that the fakeroot/
fakechroot combo can have the right libraries available.