spire: improve systemd service verification order
The prometheus metric 'node_systemd_system_running' reports the state
of 'systemctl is-system-running', which requires that no units have
failed.

As such, this will only succeed if all of our other checks succeed. It
provides less helpful information than a service-by-service list of
which services are wrong. So move it to the end of the function, after
we do the checks that would give us more detailed reporting.