debian-iso: bump installation CD to buster
This updates the script used to validate new versions of the debian ISO
provided by upstream, along with the current version specification.

This update is part of the overall upgrade from stretch to buster.
Without it, the version of CRI-O we compile will crash due to the newer
libc being missing.