spire: wait until the supervisor is accessible
Because we immediately would jump into issuing commands to the
supervisor, it was possible that we would do so slightly before the
SSH server was actually online, which would cause the install to abort.

Introduce an explicit wait so that we don't proceed until the server is
ready for us to do so.