spire: pause for supervisor ssh before key scan
We pull down the host key for the supervisor via a process involving
ssh-keyscan and a set of fingerprints scraped from the MOTD of the
supervisor. For some reason, on Buster, ssh isn't always running at the
point the MOTD is printed -- so we don't always find the SSH hostkeys
we need.

Add a loop to wait for the SSH server to be up before continuing with
the pull process.