spire: make modprobe work with mismatched kernel
When we're using a buster deploy-chroot with a non-buster host, it's
likely that the kernel versions might not match, which would lead to a
failure to run modprobe in the chroot. When we actually do need to run
modprobe, there's not much we can do, but when we don't actually need
to run the command, we can easily bypass it.