build-chroot: fix gpg keys for different versions
When we have different versions of GPG inside and outside of the build
chroot, which can occur when the host is stretch and the chroot is
buster, the //upload target may fail to sign packages through reprepro.
This is because the keys transferred in by the host might not be
compatible with the new gpg version; this can be fixed by running any
gpg command, so run gpg --list-keys on entry.

This was confusing to debug, because the error was a "key not found"
error, and if you then ran gpg --list-keys to check, you'd see it did
actually exist, and then if you ran the upload again, it would
succeed... but only until the next time you reopened the chroot.