spire: add qemu debug option to virt VM launches
It is frequently useful for debugging to have access to the exact
invocations of the qemu instances that spire executes, but learning
that information usually requires modifying the code to add debugging
statements, or carefully extracting the info from /proc/<pid>/cmdline.

Add an option to print the command lines before the VMs are launched,
in a form that can be immediately executed in the debug shell.