spire: fix disabling user-grant service
Previously, the conditionals for user-grant were ordered such that, if
there were the appropriate temporary credentials, it would attempt to
verify that the service worked, even if it was disabled in the
configuration. (This could happen if, for example, you were using spire
with user-grant enabled, and then disabled it.)

Reorder the conditionals so that user-grant is not verified in this
case, because there's no way for verification to succeed.