bazel: force host python version to python 2
Starting in bazel 0.27, the host python version is forced to python 3.
Unfortunately, containerregistry (a dependency of rules_bazel) is not
compatible with python 3, so we have to pin bazel back to python 2.

We should fix this once containerregistry updates, so that we aren't
stuck on an old version of python.