bazel: eliminate zip_file rule
Now that we no longer need zip_file to build spire, we can remove it
from our repository. We've been having to carry around this custom code
for a while, and it's been a consistent maintenance burden to deal
with. Eliminating it is a serious relief.

There is an upstream version in rules_pkg now, but it is incomplete; it
doesn't support path remapping, which we would require. But now we
don't need it at all, which is even better.