spire: set up python resource packages for deps
In order to transition spire to build using py_binary, we're going to
need to have all of spire's dependencies be accessible as python
packages.

Set up a helper macro to instantiate a stub __init__.py and create a
filegroup that can be later incorporated.

We don't use a py_library, because we have packages that include "-"
characters in them, and py_library will error out on that, even though
python itself will let us import them. (Just not with the import
statement.)