debian-iso: add autoupdate script
Because we want to spend less time working out how to update our
upstream dependencies, add a script that can automatically determine
the latest debian ISO, and automatically make a commit to upgrade to
it.

This is as simple as the following command:

    $ ./bump-debian-iso.sh auto

It includes a basic general-purpose bazel code modification tool that
should help us make these kinds of autoupdate scripts in general.