rook: add basic build of binaries and container
In order to manage our disk cluster, we need a build of rook. Our
software supply chain standards mean that we have to build it
ourselves.

We also need to pull in numerous Rook dependencies, update the version
of at least one dependency, build tini, and package up Rook in a usable
OCI.